requirements.txt
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.3
Pillow==8.4.0
pyparsing==3.0.4
python-dateutil==2.8.2
six==1.16.0
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
Werkzeug==2.0.2
